Updated create iriscodes from saved with significantly less warnings
The algorithm to find the edge after segmentation is now fixed, but still fails in some cases. This is sometimes because the eyelid detection cuts the pupil off and for other unclear reasons. Nevertheless, a lot of eyes now work and it is creating iris codes now takes much less time since we use the saved outputs from the segmentation.